refactor: 공통 KakaoMapView 컴포저블을 presentation으로 이동 (#551)
지도를 쓰는 화면들이 공유할 수 있도록 matching 에 있던 KakaoMapView 를
presentation 모듈(com.project200.presentation.compose.map)로 옮긴다.
presentation 에 libs.kakao.map 의존성을 추가하고, MatchingMapScreen 은
이동한 KakaoMapView 를 import 한다.
